test(distributed): add Ulysses CP correctness tests
Rename test_ring_attn_cp.py to test_cp.py now that the file covers both CP
strategies, and hoist the QKV-builder and shard-comparison helpers out of
TestRingAttnCP to module level so TestUlyssesCP can reuse them. _build_inputs
takes nheads_k so the GQA case can request fewer KV heads than the CP size.

TestUlyssesCP asserts bitwise equality against the single-GPU reference for
FA2/FA3/FA4, plus a nheads_k=1 case that exercises _replicate_kv_heads. The
all-to-all only moves data between ranks, so no rounding-order difference is
introduced and exact equality is the right assertion.
